requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
qdrant-client==1.6.9
psycopg2-binary==2.9.9
boto3==1.33.6
transformers==4.35.2
Pillow==10.1.0
numpy==1.26.2
torch==2.1.1
torchvision==0.16.1
